requirements.txt
disnake
python-dotenv
chat-exporter